requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
discord.py==1.7.3
google==3.0.0
google-api-core==1.30.0
google-api-python-client==1.12.8
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-images-download==2.8.0
Google-Images-Search==1.3.8
googleapis-common-protos==1.53.0
googlesearch-python==1.0.1
httplib2==0.19.1
idna==2.10
multidict==5.1.0
packaging==20.9
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfiglet==0.8.post1
PyNaCl==1.4.0
pyparsing==2.4.7
python-dotenv==0.17.1
python-resize-image==1.1.19
pytz==2021.1
requests==2.25.1
rsa==4.7.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
termcolor==1.1.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
yarl==1.6.3
youtube-dl==2021.6.6